chore: move auth routes into specific router module
