+++ commands/serve/web/components/Hub/SelectEngine/SelectEngine.jsx
@@ -19,12 +19,12 @@ const SelectEngine = () => {
 
     // DEAUTHORIZE PREVIOUS AUTHENTICATION
     const handleAuthCheck = async () => {
-      // const { isAuthorized } = await (await fetch('/isAuthorized')).json();
+      // const { isAuthorized } = await (await fetch('/auth/isAuthorized')).json();
       // console.log('[auth_check_selectEngine]: is AUTH', isAuthorized);
       // if (isAuthorized) {
       // console.log('[auth_check_selectEngine]: DEAUTH EFFECT!');
       try {
-        await (await fetch('/deauthorize')).json();
+        await (await fetch('/auth/deauthorize')).json();
         console.log('[auth_check_selectEngine]: DEAUTH OK');
       } catch (error) {
         console.log('[auth_check_selectEngine]: DEAUTH ERROR', error);
