+++ commands/serve/web/connect.js
@@ -119,9 +119,9 @@ const connect = async () => {
     } = await getConnectionInfo();
 
     // if no clientId + user is already authorized -> deAuthorize user
-    const { isAuthorized } = await (await fetch('/isAuthorized')).json();
+    const { isAuthorized } = await (await fetch('/auth/isAuthorized')).json();
     if (!clientId && isAuthorized) {
-      await (await fetch('/deauthorize')).json();
+      await (await fetch('/auth/deauthorize')).json();
     }
 
     if (webIntegrationId) {
@@ -142,7 +142,7 @@ const connect = async () => {
     if (clientId) {
       return {
         getDocList: async () => {
-          const URL = `/oauth?host=${host}&clientId=${clientId}`;
+          const URL = `/auth/oauth?host=${host}&clientId=${clientId}`;
           const resp = await (await fetch(URL)).json();
           if (resp.redirectUrl) window.location.href = resp.redirectUrl;
         },
@@ -175,7 +175,7 @@ const openApp = async (id) => {
       const authInstance = await getAuthInstance({ webIntegrationId, host });
       url = await authInstance.generateWebsocketUrl(id);
     } else if (clientId) {
-      const { webSocketUrl } = await (await fetch(`/getSocketUrl/${id}`)).json();
+      const { webSocketUrl } = await (await fetch(`/auth/getSocketUrl/${id}`)).json();
       url = webSocketUrl;
     } else {
       url = SenseUtilities.buildUrl(enigmaInfo);
